Added profile.html and profile.controller.js
I added the profile page stuff I had showed you yesterday. The buttons are working now! All that is really left I think is pulling from and manipulating firebase. Please note that I added a button to login.html just to be able to connect to the profile page as well as a function in login.controller.html to be able to redirect to the profile page. I will, of course, remove these once we are able to reach the profile page in an orderly fashion.