Fixing the Website Look
For now the changes are temporary but wanted to get some stuff done for the presentation.
If time allows will try to add some content to a few sections just to show a more cleaner idea of what we planned to do.
Tried to get windowsLoca storage to work but it still does not work :(